Rename compound.py to sphinxrenderer.py
The module contains SphinxRenderer which handles all types of nodes
rather than compound nodes only.
