+++ tests/test_utils.py
@@ -2,7 +2,7 @@
 from unittest import TestCase
 from xml.dom import minidom
 
-from breathe.renderer.compound import get_param_decl, get_definition_without_template_args
+from breathe.renderer.sphinxrenderer import get_param_decl, get_definition_without_template_args
 from breathe.parser.compoundsuper import memberdefType
 from breathe.directives import PathHandler
